finalized the inclusion of 128 bit refinement

+++ src/base/global.c
@@ -56,8 +56,8 @@ double tot_nissa_comm_time=2;
 
 //verbosity
 int verb_call=0;
-const double nissa_default_verbosity=2;
-double nissa_verbosity=2;
+const int nissa_default_verbosity=2;
+int nissa_verbosity=2;
 
 //128 bit precision
 const int nissa_default_use_128_bit_precision=0;
